Enum gdk_sys::GdkEventType

pub enum GdkEventType {
    Nothing,
    Delete,
    Destroy,
    Expose,
    MotionNotify,
    ButtonPress,
    DoubleButtonPress,
    TripleButtonPress,
    ButtonRelease,
    KeyPress,
    KeyRelease,
    EnterNotify,
    LeaveNotify,
    FocusChange,
    Configure,
    Map,
    Unmap,
    PropertyNotify,
    SelectionClear,
    SelectionRequest,
    SelectionNotify,
    ProximityIn,
    ProximityOut,
    DragEnter,
    DragLeave,
    DragMotion,
    DragStatus,
    DropStart,
    DropFinished,
    ClientEvent,
    VisibilityNotify,
    Scroll,
    WindowState,
    Setting,
    OwnerChange,
    GrabBroken,
    Damage,
    TouchBegin,
    TouchUpdate,
    TouchEnd,
    TouchCancel,
}

Specifies the type of the event.

Do not confuse these events with the signals that GTK+ widgets emit. Although many of these events result in corresponding signals being emitted, the events are often transformed or filtered along the way.

In some language bindings, the values EventType::2buttonPress and EventType::3buttonPress would translate into something syntactically invalid (eg Gdk.EventType.2ButtonPress, where a symbol is not allowed to start with a number). In that case, the aliases EventType::DoubleButtonPress and EventType::TripleButtonPress can be used instead.

Variants

Nothing

a special code to indicate a null event.

Delete

the window manager has requested that the toplevel window be hidden or destroyed, usually when the user clicks on a special icon in the title bar.

Destroy

the window has been destroyed.

Expose

all or part of the window has become visible and needs to be redrawn.

MotionNotify

the pointer (usually a mouse) has moved.

ButtonPress

a mouse button has been pressed.

DoubleButtonPress

alias for EventType::2buttonPress, added in 3.6.

TripleButtonPress

alias for EventType::3buttonPress, added in 3.6.

ButtonRelease

a mouse button has been released.

KeyPress

a key has been pressed.

KeyRelease

a key has been released.

EnterNotify

the pointer has entered the window.

LeaveNotify

the pointer has left the window.

FocusChange

the keyboard focus has entered or left the window.

Configure

the size, position or stacking order of the window has changed. Note that GTK+ discards these events for WindowType::Child windows.

Map

the window has been mapped.

Unmap

the window has been unmapped.

PropertyNotify

a property on the window has been changed or deleted.

SelectionClear

the application has lost ownership of a selection.

SelectionRequest

another application has requested a selection.

SelectionNotify

a selection has been received.

ProximityIn

an input device has moved into contact with a sensing surface (e.g. a touchscreen or graphics tablet).

ProximityOut

an input device has moved out of contact with a sensing surface.

DragEnter

the mouse has entered the window while a drag is in progress.

DragLeave

the mouse has left the window while a drag is in progress.

DragMotion

the mouse has moved in the window while a drag is in progress.

DragStatus

the status of the drag operation initiated by the window has changed.

DropStart

a drop operation onto the window has started.

DropFinished

the drop operation initiated by the window has completed.

ClientEvent

a message has been received from another application.

VisibilityNotify

the window visibility status has changed.

Scroll

the scroll wheel was turned

WindowState

the state of a window has changed. See WindowState for the possible window states

Setting

a setting has been modified.

OwnerChange

the owner of a selection has changed. This event type was added in 2.6

GrabBroken

a pointer or keyboard grab was broken. This event type was added in 2.8.

Damage

the content of the window has been changed. This event type was added in 2.14.

TouchBegin

A new touch event sequence has just started. This event type was added in 3.4.

TouchUpdate

A touch event sequence has been updated. This event type was added in 3.4.

TouchEnd

A touch event sequence has finished. This event type was added in 3.4.

TouchCancel

A touch event sequence has been canceled. This event type was added in 3.4.