Enum gdk_sys::GdkWindowTypeHint

pub enum GdkWindowTypeHint {
    Normal,
    Dialog,
    Menu,
    Toolbar,
    Splashscreen,
    Utility,
    Dock,
    Desktop,
    DropdownMenu,
    PopupMenu,
    Tooltip,
    Notification,
    Combo,
    Dnd,
}

These are hints for the window manager that indicate what type of function the window has. The window manager can use this when determining decoration and behaviour of the window. The hint must be set before mapping the window.

See the Extended Window Manager Hints specification for more details about window types.

Variants

Normal

Normal toplevel window.

Dialog

Dialog window.

Menu

Window used to implement a menu; GTK+ uses this hint only for torn-off menus, see GtkTearoffMenuItem.

Toolbar

Window used to implement toolbars.

Splashscreen

Window used to display a splash screen during application startup.

Utility

Utility windows which are not detached toolbars or dialogs.

Dock

Used for creating dock or panel windows.

Desktop

Used for creating the desktop background window.

DropdownMenu

A menu that belongs to a menubar.

PopupMenu

A menu that does not belong to a menubar, e.g. a context menu.

Tooltip

A tooltip.

Notification

A notification - typically a “bubble” that belongs to a status icon.

Combo

A popup from a combo box.

Dnd

A window that is used to implement a DND cursor.